change ConnectWalletButtonV2 to ConnectWalletButton; remove old ConnectWalletButton
