+++ components/Button/ConnectWalletButton.tsx
@@ -1,88 +1,70 @@
-import Button, { ButtonProps } from "./index";
 import { useMemo, useState } from "react";
 import SignupModal from "@/components/Modal/SignupModal";
 import WalletConnectModal from "@/components/Modal/WalletConnectModal";
 import { useAccount } from "wagmi";
 import SignConnectMessageModal from "@/components/Modal/SignConnectMessageModal";
 import useAuthStore from "@/store/auth/store";
-import { useRouter } from "next/navigation";
+import Button from "@/components/Button/index";
+import { useAuth } from "@/hooks/useAuth";
 
-interface Props extends ButtonProps {
+interface Props {
+  className?: string;
   children?: React.ReactNode;
-  mode?: "link" | "modal";
+  action?: (accessToken?: string) => void;
+  showConnectButton?: boolean;
 }
 
 export default function ConnectWalletButton({
+  action,
   className,
-  mode = "modal",
+  showConnectButton,
   children,
-  ...rest
 }: Props) {
-  const router = useRouter();
-  const { isConnected, address } = useAccount();
   const [showWalletConnect, setShowWalletConnect] = useState(false);
   const [showSignMessage, setShowSignMessage] = useState(false);
   const [showSignup, setShowSignup] = useState(false);
-  const acceptedTerms = useAuthStore((state) => state.profile?.acceptedTerms);
-  const accessToken = useAuthStore((state) => state.credentials?.accessToken);
-  const expiredDate = useAuthStore(
-    (state) => state.credentials?.accessTokenExpire,
-  );
-  const userWallet = useAuthStore((state) => state.profile?.publicKey);
-  const isCorrectWallet = useMemo(() => {
-    if (!userWallet || !address) return false;
-    return userWallet.toLowerCase() === address.toLowerCase();
-  }, [userWallet, address]);
-  const isExpired = useMemo(() => {
-    return !!expiredDate && expiredDate < Date.now();
-  }, [expiredDate]);
+
+  const { isValidSession } = useAuth();
 
   const handleConnectWallet = () => {
-    if (mode === "modal") {
+    if (!isValidSession) {
       setShowWalletConnect(true);
     } else {
-      router.push("/connect");
+      // Access Token has been saved in auth store
+      action?.();
     }
   };
 
-  // if ((isConnected || mode === 'link') && acceptedTerms && !!accessToken && !isExpired && isCorrectWallet) {
-  //   return children
-  // }
-
   return (
     <>
-      {(isConnected || mode === "link") &&
-      acceptedTerms &&
-      !!accessToken &&
-      !isExpired &&
-      isCorrectWallet ? (
-        children
-      ) : (
-        <>
-          <Button
-            type="button"
-            onClick={handleConnectWallet}
-            className={className}
-            {...rest}
-          >
+      <div className={className} onClick={handleConnectWallet}>
+        {showConnectButton && !isValidSession ? (
+          <Button type="button" onClick={handleConnectWallet}>
             Connect Wallet
           </Button>
+        ) : (
+          children
+        )}
+      </div>
 
-          <WalletConnectModal
-            show={showWalletConnect}
-            onSignMessage={() => setShowSignMessage(true)}
-            onClose={() => setShowWalletConnect(false)}
-          />
+      <WalletConnectModal
+        show={showWalletConnect}
+        onSignMessage={() => setShowSignMessage(true)}
+        onClose={() => setShowWalletConnect(false)}
+      />
 
-          <SignConnectMessageModal
-            show={showSignMessage}
-            onSignup={() => setShowSignup(true)}
-            onClose={() => setShowSignMessage(false)}
-          />
+      <SignConnectMessageModal
+        show={showSignMessage}
+        onConnectSuccess={(accessToken) => action?.(accessToken)}
+        onSignup={() => setShowSignup(true)}
+        onClose={() => setShowSignMessage(false)}
+      />
 
-          <SignupModal show={showSignup} onClose={() => setShowSignup(false)} />
-        </>
-      )}
+      <SignupModal
+        show={showSignup}
+        onClose={() => setShowSignup(false)}
+        onSignupSuccess={(accessToken) => action?.(accessToken)}
+      />
     </>
   );
 }
