+++ components/Layout/MainHeader/index.tsx
@@ -4,7 +4,7 @@ import Image from "next/image";
 import { useState } from "react";
 import Icon from "@/components/Icon";
 import Dropdown from "@/components/Dropdown";
-import ConnectWalletButton from "@/components/Button/ConnectWalletButtonV2";
+import ConnectWalletButton from "@/components/Button/ConnectWalletButton";
 import { navs } from "@/config/nav";
 import useAuthStore from "@/store/auth/store";
 import SearchInput from "@/components/Layout/MainHeader/SearchInput";
