+++ components/Pages/Launchpad/ProjectDetails/LauchpadBanner/RoundActions/RoundActionMinting.tsx
@@ -217,7 +217,7 @@ export default function RoundActionMinting({
           </div>
         )}
         <div className="flex-1 w-full">
-          <ConnectWalletButton scale="lg" className="w-full">
+          <ConnectWalletButton className="w-full">
             <Button
               disabled={
                 // roundType == '2' &&
