+++ components/Pages/Launchpad/ProjectDetails/LauchpadBanner/RoundActions/RoundActionUpcoming.tsx
@@ -56,7 +56,7 @@ export default function RoundActionUpcoming({
 
       {round.type === "U2UPremintRoundZero" ||
       round.type === "U2UMintRoundZero" ? (
-        <ConnectWalletButton scale="lg" className="w-full">
+        <ConnectWalletButton className="w-full">
           {!isSubscribed ? (
             <Button
               scale="lg"
@@ -107,7 +107,7 @@ export default function RoundActionUpcoming({
           )}
         </div>
       ) : (
-        <ConnectWalletButton scale="lg" className="w-full">
+        <ConnectWalletButton className="w-full">
           <Button disabled scale="lg" className="w-full">
             Mint now
           </Button>
